+++ packages/runtime/src/runtime/customWidget.tsx
@@ -1,10 +1,15 @@
 import {
   CustomScopeProvider,
+  error,
   useRegisterBindings,
 } from "@ensembleui/react-framework";
-import type { CustomWidgetModel } from "@ensembleui/react-framework";
-import React from "react";
+import type {
+  CustomWidgetModel,
+  EnsembleAction,
+} from "@ensembleui/react-framework";
+import React, { useEffect, useState } from "react";
 import { EnsembleRuntime } from "./runtime";
+import { useEnsembleAction } from "./hooks/useEnsembleAction";
 
 export interface CustomWidgetProps {
   inputs: Record<string, unknown>;
@@ -19,9 +24,35 @@ export const createCustomWidget = (
     });
     return (
       <CustomScopeProvider value={values ?? inputs}>
-        {EnsembleRuntime.render([widget.body])}
+        <OnLoadAction action={widget?.onLoad} context={values ?? inputs}>
+          {EnsembleRuntime.render([widget.body])}
+        </OnLoadAction>
       </CustomScopeProvider>
     );
   };
   return CustomWidget;
+};
+
+const OnLoadAction: React.FC<
+  React.PropsWithChildren<{
+    action?: EnsembleAction;
+    context: Record<string, unknown>;
+  }>
+> = ({ action, children, context }) => {
+  const onLoadAction = useEnsembleAction(action);
+  const [isComplete, setIsComplete] = useState(false);
+  useEffect(() => {
+    if (!onLoadAction?.callback || isComplete) {
+      return;
+    }
+    try {
+      onLoadAction.callback(context);
+    } catch (e) {
+      error(e);
+    } finally {
+      setIsComplete(true);
+    }
+  }, [context, isComplete, onLoadAction, onLoadAction?.callback]);
+
+  return <>{children}</>;
 };