+++ apps/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ensembleui/react-kitchen-sink
 
+## 0.0.12
+
+### Patch Changes
+
+- b9633f3: added stop event propogation in button onTap action
+- db41e4a: Added support of default value in dropdown
+- 2196d64: exposed url params and route query into screen context
+- 208025e: add getDaysDifference, getMonthsDifference and getYearsDifference in date formatter api
+- Updated dependencies [b9633f3]
+- Updated dependencies [db41e4a]
+- Updated dependencies [2196d64]
+- Updated dependencies [208025e]
+  - @ensembleui/react-runtime@0.1.21
+  - @ensembleui/react-framework@0.1.17
+
+## 0.0.11
+
+### Patch Changes
+
+- 6cf809f: make showDialog screen aware
+- 6cf809f: Various fixes related to scoping
+- Updated dependencies [6cf809f]
+- Updated dependencies [6cf809f]
+  - @ensembleui/react-runtime@0.1.20
+  - @ensembleui/react-framework@0.1.16
+
 ## 0.0.10
 
 ### Patch Changes