+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ensembleui/react-framework
 
+## 0.1.17
+
+### Patch Changes
+
+- 208025e: add getDaysDifference, getMonthsDifference and getYearsDifference in date formatter api
+
+## 0.1.16
+
+### Patch Changes
+
+- 6cf809f: Various fixes related to scoping
+
 ## 0.1.15
 
 ### Patch Changes