+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ensembleui/react-runtime
 
+## 0.1.21
+
+### Patch Changes
+
+- b9633f3: added stop event propogation in button onTap action
+- db41e4a: Added support of default value in dropdown
+- 2196d64: exposed url params and route query into screen context
+- Updated dependencies [208025e]
+  - @ensembleui/react-framework@0.1.17
+
+## 0.1.20
+
+### Patch Changes
+
+- 6cf809f: make showDialog screen aware
+- 6cf809f: Various fixes related to scoping
+- Updated dependencies [6cf809f]
+  - @ensembleui/react-framework@0.1.16
+
 ## 0.1.19
 
 ### Patch Changes